WIP: Motor library

+++ encoder.cpp
@@ -11,8 +11,8 @@ void Encoder::init(int pin){
 }
 
 
-float Encoder::getRPM(void){
-    return _RPM;
+uint16_t Encoder::getRPM(void){
+    return (uint16_t) _RPM;
 }
 
 
@@ -31,7 +31,7 @@ void Encoder::timerInterrupt(void){
     
     _micros += 50; //ISR is called each 50us
     if(((_micros*1000) - _RPM_read_timeout) > ENCODER_RPM_QUERY_INTERVAL){
-        _RPM = (float) (_RPM_counter * 60 * 1000) / (ENCODER_HOLES * ENCODER_RPM_QUERY_INTERVAL);
+        _RPM = (_RPM_counter * 60 * 1000) / (ENCODER_HOLES * ENCODER_RPM_QUERY_INTERVAL);
         _RPM_counter = 0;
         
         _RPM_read_timeout = (_micros*1000) + ENCODER_RPM_QUERY_INTERVAL;
